refactor: move the event closed hero from the apps to components package
LIIKUNTA-474
