--- apps/sports-helsinki/src/domain/event/eventClosedHero/eventClosedHero.module.scss
@@ -1,32 +1,0 @@
-@import 'src/styles/breakpoints';
-
-.eventClosedHero {
-  background: var(--color-theme-background);
-  height: 100%;
-  padding: 5.625rem var(--spacing-m) 10rem;
-  text-align: center;
-  box-sizing: border-box;
-
-  @include respond_above(s) {
-    padding-top: 10.375rem;
-    padding-bottom: 10.375rem;
-  }
-
-  h1 {
-    max-width: 700px;
-    margin-left: auto;
-    margin-right: auto;
-  }
-
-  p {
-    font-size: 1.25rem;
-    max-width: 610px;
-    margin: 0 auto 2rem;
-    line-height: 2rem;
-  }
-
-  button {
-    padding-left: 3.5rem;
-    padding-right: 3.5rem;
-  }
-}
