+++ packages/components/src/components/eventClosedHero/eventClosedHero.module.scss
