+++ packages/components/src/components/index.ts
@@ -61,3 +61,4 @@ export { default as LargeEventCard } from './eventCard/LargeEventCard';
 export { default as EventName } from './eventName/EventName';
 export { default as EventKeywords } from './eventKeywords/EventKeywords';
 export { default as EventList } from './eventList/EventList';
+export { default as EventClosedHero } from './eventClosedHero/EventClosedHero';
